Moves CSSStyleDeclaration to mocks dir
