+++ lib/StylePropertyMapReadOnly.js
@@ -75,6 +75,10 @@ export default class StylePropertyMapReadOnly {
    * @param {string} propertyName
    * @returns {CSSStyleValue|null}
    * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/StylePropertyMapReadOnly/get}
+   *
+   * Note: We may want to change the behavior based on this from the spec
+   * "In practice, since the values are "hidden" behind a .get() method call,
+   * UAs can delay computing anything until a given property is actually requested."
    */
   get(propertyName) {
     return this.#styles[propertyName] || null;
