+++ lib/StylePropertyMapReadOnly.test.js
@@ -1,34 +1,7 @@
 import { describe, it } from "node:test";
 import assert from "node:assert/strict";
+import MockCSSStyleDeclaration from "./mocks/MockCSSStyleDeclaration.js";
 import StylePropertyMapReadOnly from "./StylePropertyMapReadOnly.js";
-
-/**
- * In browsers, CSSStyleDeclaration will be available so we can use props and
- * methods of it in StylePropertyMapReadOnly, but in these tests, it won't
- * exist so we need to mock just the parts we need.
- */
-class MockCSSStyleDeclaration {
-  constructor(styles) {
-    const props = Object.keys(styles);
-    this.props = props;
-    this.styles = props.reduce((obj, key) => {
-      obj[key] = styles[key];
-      return obj;
-    }, {});
-  }
-
-  get length() {
-    return this.props.length;
-  }
-
-  item(index) {
-    return this.props[index];
-  }
-
-  getPropertyValue(propertyName) {
-    return this.styles[propertyName];
-  }
-}
 
 describe("class StylePropertyMapReadOnly", () => {
   const mockStyles = {
